requirements.txt
flask
mysql-connector-python
python-dotenv
openai
numpy
flask-socketio